Optimized runtime performance of the SAM validation algorithm by switching to a statically generated SAM CFN specification file that should be manually updated via the `sam2CFN` utility.
